put fold & try fold in their own module
