+++ src/multi/tests.rs
@@ -12,8 +12,8 @@ use crate::{
 use crate::{
   lib::std::vec::Vec,
   multi::{
-    count, fold, fold_many0, fold_many1, fold_many_m_n, length_count, many0, many1, many_m_n,
-    many_till, separated_list0, separated_list1,
+    count, fold_many0, fold_many1, fold_many_m_n, length_count, many0, many1, many_m_n, many_till,
+    separated_list0, separated_list1,
   },
 };
 
@@ -539,196 +539,6 @@ fn many1_count_test() {
     Err(Err::Error(error_position!(
       &b"hello"[..],
       ErrorKind::Many1Count
-    )))
-  );
-}
-
-#[test]
-#[cfg(feature = "alloc")]
-fn fold_test() {
-  fn fold_into_vec<T>(mut acc: Vec<T>, item: T) -> Vec<T> {
-    acc.push(item);
-    acc
-  }
-
-  // should not go into an infinite loop
-  fn fold_error_0(i: &[u8]) -> IResult<&[u8], Vec<&[u8]>> {
-    fn tst(input: &[u8]) -> IResult<&[u8], &[u8]> {
-      Err(Err::Error(error_position!(input, ErrorKind::Tag)))
-    }
-    fold(0.., tst, Vec::new, fold_into_vec)(i)
-  }
-
-  let a = &b"abcdef"[..];
-  assert_eq!(fold_error_0(a), Ok((a, Vec::new())));
-
-  fn fold_error_1(i: &[u8]) -> IResult<&[u8], Vec<&[u8]>> {
-    fn tst(input: &[u8]) -> IResult<&[u8], &[u8]> {
-      Err(Err::Error(error_position!(input, ErrorKind::Tag)))
-    }
-    fold(1.., tst, Vec::new, fold_into_vec)(i)
-  }
-
-  let a = &b"abcdef"[..];
-  assert_eq!(
-    fold_error_1(a),
-    Err(Err::Error(error_position!(a, ErrorKind::Tag)))
-  );
-
-  fn fold_error(i: &[u8]) -> IResult<&[u8], Vec<&[u8]>> {
-    fold(0.., tag(""), Vec::new, fold_into_vec)(i)
-  }
-
-  let a = &b"abcdef"[..];
-  assert_eq!(
-    fold_error(a),
-    Err(Err::Error(error_position!(a, ErrorKind::Fold)))
-  );
-
-  fn fold_invalid(i: &[u8]) -> IResult<&[u8], Vec<&[u8]>> {
-    fold(2..=1, tag("a"), Vec::new, fold_into_vec)(i)
-  }
-
-  let a = &b"a"[..];
-  let b = &b"b"[..];
-  assert_eq!(
-    fold_invalid(a),
-    Err(Err::Failure(error_position!(a, ErrorKind::Fold)))
-  );
-  assert_eq!(
-    fold_invalid(b),
-    Err(Err::Failure(error_position!(b, ErrorKind::Fold)))
-  );
-
-  fn fold_any(i: &[u8]) -> IResult<&[u8], Vec<&[u8]>> {
-    fold(0.., tag("abcd"), Vec::new, fold_into_vec)(i)
-  }
-
-  let a = &b"abcdef"[..];
-  let b = &b"abcdabcdefgh"[..];
-  let c = &b"azerty"[..];
-  let d = &b"abcdab"[..];
-  let e = &b"abcd"[..];
-  let f = &b""[..];
-  assert_eq!(fold_any(a), Ok((&b"ef"[..], vec![&b"abcd"[..]])));
-  assert_eq!(
-    fold_any(b),
-    Ok((&b"efgh"[..], vec![&b"abcd"[..], &b"abcd"[..]]))
-  );
-  assert_eq!(fold_any(c), Ok((c, Vec::new())));
-  assert_eq!(fold_any(d), Err(Err::Incomplete(Needed::new(2))));
-  assert_eq!(fold_any(e), Err(Err::Incomplete(Needed::new(4))));
-  assert_eq!(fold_any(f), Err(Err::Incomplete(Needed::new(4))));
-
-  fn fold_1(i: &[u8]) -> IResult<&[u8], Vec<&[u8]>> {
-    fold(1.., tag("abcd"), Vec::new, fold_into_vec)(i)
-  }
-
-  let a = &b"abcdef"[..];
-  let b = &b"abcdabcdefgh"[..];
-  let c = &b"azerty"[..];
-  let d = &b"abcdab"[..];
-  let res1 = vec![&b"abcd"[..]];
-  assert_eq!(fold_1(a), Ok((&b"ef"[..], res1)));
-  let res2 = vec![&b"abcd"[..], &b"abcd"[..]];
-  assert_eq!(fold_1(b), Ok((&b"efgh"[..], res2)));
-  assert_eq!(
-    fold_1(c),
-    Err(Err::Error(error_position!(c, ErrorKind::Tag)))
-  );
-  assert_eq!(fold_1(d), Err(Err::Incomplete(Needed::new(2))));
-
-  fn fold_m_n(i: &[u8]) -> IResult<&[u8], Vec<&[u8]>> {
-    fold(2..=3, tag("Abcd"), Vec::new, fold_into_vec)(i)
-  }
-
-  let a = &b"Abcdef"[..];
-  let b = &b"AbcdAbcdefgh"[..];
-  let c = &b"AbcdAbcdAbcdAbcdefgh"[..];
-  let d = &b"AbcdAbcdAbcdAbcdAbcdefgh"[..];
-  let e = &b"AbcdAb"[..];
-  assert_eq!(
-    fold_m_n(a),
-    Err(Err::Error(error_position!(&b"ef"[..], ErrorKind::Tag)))
-  );
-  let res1 = vec![&b"Abcd"[..], &b"Abcd"[..]];
-  assert_eq!(fold_m_n(b), Ok((&b"efgh"[..], res1)));
-  let res2 = vec![&b"Abcd"[..], &b"Abcd"[..], &b"Abcd"[..], &b"Abcd"[..]];
-  assert_eq!(fold_m_n(c), Ok((&b"efgh"[..], res2)));
-  let res3 = vec![&b"Abcd"[..], &b"Abcd"[..], &b"Abcd"[..], &b"Abcd"[..]];
-  assert_eq!(fold_m_n(d), Ok((&b"Abcdefgh"[..], res3)));
-  assert_eq!(fold_m_n(e), Err(Err::Incomplete(Needed::new(2))));
-
-  fn fold_fixed(i: &[u8]) -> IResult<&[u8], Vec<&[u8]>> {
-    fold(2..2, tag("Abcd"), Vec::new, fold_into_vec)(i)
-  }
-
-  let a = &b"Abcdef"[..];
-  let b = &b"AbcdAbcdefgh"[..];
-  let c = &b"AbcdAbcdAbcdAbcdefgh"[..];
-  let d = &b"AbcdAb"[..];
-  assert_eq!(
-    fold_fixed(a),
-    Err(Err::Error(error_position!(&b"ef"[..], ErrorKind::Tag)))
-  );
-  let res1 = vec![&b"Abcd"[..], &b"Abcd"[..]];
-  assert_eq!(fold_fixed(b), Ok((&b"efgh"[..], res1)));
-  let res2 = vec![&b"Abcd"[..], &b"Abcd"[..]];
-  assert_eq!(fold_fixed(c), Ok((&b"AbcdAbcdefgh"[..], res2)));
-  assert_eq!(fold_fixed(d), Err(Err::Incomplete(Needed::new(2))));
-
-  fn fold_exclusive(i: &[u8]) -> IResult<&[u8], Vec<&[u8]>> {
-    fold(0..1, tag("Abcd"), Vec::new, fold_into_vec)(i)
-  }
-
-  let a = &b"AbcdAbcdAbcd"[..];
-  let b = &b"AAA"[..];
-  let res1 = vec![&b"Abcd"[..]];
-  assert_eq!(fold_exclusive(a), Ok((&b"AbcdAbcd"[..], res1)));
-  let res2 = vec![];
-  assert_eq!(fold_exclusive(b), Ok((&b"AAA"[..], res2)));
-
-  fn fold_never(i: &[u8]) -> IResult<&[u8], Vec<&[u8]>> {
-    fold(0..=0, tag("A"), Vec::new, fold_into_vec)(i)
-  }
-
-  let a = &b"AAA"[..];
-  let b = &b"B"[..];
-  let res1 = vec![&b"A"[..]];
-  assert_eq!(fold_never(a), Ok((&b"AA"[..], res1)));
-  assert_eq!(fold_never(b), Ok((&b"B"[..], Vec::new())));
-
-  fn fold_usize(i: &[u8]) -> IResult<&[u8], Vec<&[u8]>> {
-    fold(1, tag("A"), Vec::new, fold_into_vec)(i)
-  }
-
-  let a = &b"AAA"[..];
-  let res1 = vec![&b"A"[..], &b"A"[..]];
-  assert_eq!(fold_usize(a), Ok((&b"A"[..], res1)));
-}
-
-#[test]
-fn try_fold_test() {
-  use super::try_fold;
-  use crate::{character::complete::anychar, error::Error};
-
-  let input = "it's my life";
-  let result = try_fold(
-    ..,
-    anychar,
-    || vec![42],
-    |_, _| {
-      Err(Err::Error(Error::from_error_kind(
-        input,
-        ErrorKind::TryFold,
-      )))
-    },
-  )(input);
-  assert_eq!(
-    result,
-    Err(Err::Error(Error::from_error_kind(
-      input,
-      ErrorKind::TryFold,
     )))
   );
 }
